Update client.ini
Fixed some typos and made the spelling of some titles match up with what they're listed as on the official Nintendo site. Decided not to go with how titles are displayed in the 3DS OS as most of them are missing punctuation marks, capitalized for the sake of visibility on the small screen, or have the rest of the title omitted.