Update client.ini
Fixed some typos and made the spelling of some titles match up with what they're listed as on the official Nintendo site. In the case of a few titles that were either unavailable on said site or named properly in future releases/updates on other Nintendo platforms, they're now spelled to match those and as shown on their respective developer sites.